requirements.txt
Flask==2.2.2
cassandra-driver==3.25.0
python-dotenv==1.0.1
bcrypt==4.0.1
requests==2.31.0
Werkzeug==2.2.2